fix: webembed api for imageUtils
