+++ packages/kit-bg/src/webembeds/instance/webembedApi.ts
@@ -19,8 +19,8 @@ const getOrCreateWebEmbedApiModule = memoizee(
     if (name === 'test') {
       return new (await import('../WebEmbedApiTest')).default();
     }
-    if (name === 'homeScreen') {
-      return new (await import('../WebEmbedApiHomeScreen')).default();
+    if (name === 'imageUtils') {
+      return new (await import('../WebEmbedApiImageUtils')).default();
     }
     throw new Error(
       `Unknown WebEmbed API module: ${
