+++ packages/kit-bg/src/webembeds/instance/webembedApiProxy.ts
@@ -9,7 +9,7 @@ import { RemoteApiProxyBase } from '../../apis/RemoteApiProxyBase';
 import type { IWebembedApi, IWebembedApiKeys } from './IWebembedApi';
 import type { IBackgroundApiWebembedCallMessage } from '../../apis/IBackgroundApi';
 import type WebEmbedApiChainAdaLegacy from '../WebEmbedApiChainAdaLegacy';
-import type WebEmbedApiHomeScreen from '../WebEmbedApiHomeScreen';
+import type WebEmbedApiImageUtils from '../WebEmbedApiImageUtils';
 import type WebEmbedApiSecret from '../WebEmbedApiSecret';
 import type WebEmbedApiTest from '../WebEmbedApiTest';
 
@@ -71,8 +71,8 @@ class WebembedApiProxy extends RemoteApiProxyBase implements IWebembedApi {
   secret: WebEmbedApiSecret =
     this._createProxyModule<IWebembedApiKeys>('secret');
 
-  homeScreen: WebEmbedApiHomeScreen =
-    this._createProxyModule<IWebembedApiKeys>('homeScreen');
+  imageUtils: WebEmbedApiImageUtils =
+    this._createProxyModule<IWebembedApiKeys>('imageUtils');
 }
 
 const webembedApiProxy = new WebembedApiProxy();
